refactor: 이전 PR에서 반영되지 않은 구조 반영 (#415)
